🐛 (ci): copy github/* files without rendering them
Due to the fact that github actions uses the similar style of template variables (i.e {{) cookiecutter thinks this is part of the templated file. This change introduc
es a _copy_without_render field into the cookiecutter.json. This field is mapped to a glob of `github/*` which means that everything within the `.github` folder will be copied as is. For more in
formation refer to [this](https://cookiecutter.readthedocs.io/en/latest/advanced/copy_without_render.html#copy-withou